Added Bart ARchive Tool
Bart is a tool for managing the interface between hot and cold storage.
In our case, Turbo and Dataden.

One of the design constraints for utilizing Turbo and Dataden is that
Dataden can't handle small files, so it's necessary to archive our
directories with tar before we send them to dataden. It's kind of a
pain in the ass if that compression isn't done transparently, hence the
tool.

BArT currently implements PUSH, LIST, and PULL commands. PUSH will
compress a given directory and send it to a well-known location on Turbo
where can be discovered with LIST and retrieved with PULL. PULL will
decompress the directory retrieved from archive and send it to a given
directory.